[r] Refactoring of files in src/
