+++ apis/r/src/rutilities.cpp
@@ -252,3 +252,59 @@ Rcpp::XPtr<ArrowArray> array_setup_struct(Rcpp::XPtr<ArrowArray> arrxp, int64_t 
     array->n_children = n_children;
     return arrxp;
 }
+
+
+// formerly stats.cpp
+
+//' TileDB Statistics interface
+//'
+//' The functions `tiledbsoma_stats_enable`, `tiledbsoma_stats_disable`, `tiledbsoma_stats_reset`
+//' and `tiledbsoma_stats_dump` expose the TileDB Core functionality for performance measurements
+//' and statistics.  The first three just turn on, off or reset, the fourth returns a JSON string.
+//' For convenience the function `tiledbsoma_stats_show` displays the information on the console.
+//'
+//' @export
+// [[Rcpp::export]]
+void tiledbsoma_stats_enable() {
+    tiledbsoma::stats::enable();
+}
+
+//' @rdname tiledbsoma_stats_enable
+//' @export
+// [[Rcpp::export]]
+void tiledbsoma_stats_disable() {
+    tiledbsoma::stats::disable();
+}
+
+//' @rdname tiledbsoma_stats_enable
+//' @export
+// [[Rcpp::export]]
+void tiledbsoma_stats_reset() {
+    tiledbsoma::stats::reset();
+}
+
+//' @rdname tiledbsoma_stats_enable
+//' @export
+// [[Rcpp::export]]
+std::string tiledbsoma_stats_dump() {
+    return tiledbsoma::stats::dump();
+}
+
+// formerly version.cpp
+
+//' libtiledbsoma version information
+//'
+//' Returns a string with version information for libtiledbsoma and the linked TileDB Embedded library.
+// [[Rcpp::export]]
+std::string libtiledbsoma_version() {
+    return tiledbsoma::version::as_string();
+}
+
+//' TileDB Embedded Version interface
+//'
+//' This gets the version of the TileDB Embedded library that is currently in use.
+// [[Rcpp::export]]
+Rcpp::IntegerVector tiledb_embedded_version() {
+    std::tuple<int, int, int> triple = tiledbsoma::version::embedded_version_triple();
+    return Rcpp::IntegerVector::create(std::get<0>(triple), std::get<1>(triple), std::get<2>(triple));
+}
